requirements/base.txt
# Bleeding edge Django
django==1.10.6
djangorestframework==3.6.2
Markdown==2.6.6

django-rest-swagger==2.1.2


coreapi==2.3.0
openapi-codec==1.3.1
simplejson==3.9.0

# Configuration
django-environ==0.4.0
whitenoise==3.2

# Models
django-model-utils==2.5

# Images
Pillow==3.3.0

# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
# unicode-slugify==0.1.3
# django-autoslug==1.9.3

# Time zones support
pytz==2016.6.1

# Redis support
# django-redis==4.4.4
# redis>=2.10.0


# Your custom requirements go here

django-cors-headers==2.1.0

# 2-factor auth
pyotp==2.2.6

django-choices==1.6.0

boto3==1.4.4
